+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v3.113.0
+* Added experimental `config.WithDisableOptimisticUnban` option to disable fast node unban after pessimization
+* Fixed respect start offset from `topicoptions.WithReaderGetPartitionStartOffset` for commit messages
+* Added `query.AllowImplicitSessions()` option for execute queries through `query.Client.{Exec,Query,QueryResultSet,QueryRow}` without explicit sessions
+
 ## v3.112.0
 * Added support for the `json.Unmarshaler` interface in the `CastTo` function for use in scanners, such as the `ScanStruct` method
 * Fixed the support of server-side session balancing in `database/sql` driver